+++ packages/common/src/index.ts
@@ -303,6 +303,7 @@ export type {
 } from './types/projects';
 export * from './types/promotion';
 export * from './types/queryHistory';
+export * from './types/queryHistoryList';
 export * from './types/querySources';
 export * from './types/rename';
 export * from './types/resourceViewItem';
@@ -366,6 +367,7 @@ export * from './utils/i18n/merge';
 export * from './utils/i18n/types';
 export * from './utils/item';
 export * from './utils/mergeQueryItems';
+export * from './utils/queryHistoryList';
 export * from './utils/loadLightdashProjectConfig';
 export * from './utils/lightdashSqlVariables';
 export * from './utils/metricsExplorer';